Fix spacing edge case
If the text is long and follows directly after a full-width block with the same background the text would align directly next to it. Added a padding to the top of the text to prevent this.

If two non full width media blocks with the same background and the text is taller than the image it will deviate from the standard spacings which is why I have kept it smaller than the section spacing.

Improvements on this is always appreciated.